[Refactor] dashboards.ts와 columns.ts 분리, sidemenu.ts와 dashboards.ts 중복 api 파일 병합
