+++ src/api/comment.ts
@@ -5,12 +5,14 @@ import {
   UpdateCommenttype,
   DeleteCommentParams,
 } from "../types/comments";
-//댓글 생성
+
+// 댓글 생성
 export const createComment = async (data: CreateCommentType) => {
   const response = await axiosInstance.post(apiRoutes.Comments(), data);
   return response.data;
 };
-//댓글 목록
+
+// 댓글 목록
 export async function getComments({
   cardId,
   pageParam,
@@ -31,7 +33,7 @@ export async function getComments({
   };
 }
 
-//댓글 수정
+// 댓글 수정
 export const updateComment = async (
   commentId: number,
   data: UpdateCommenttype
@@ -42,7 +44,7 @@ export const updateComment = async (
   );
   return response.data;
 };
-//댓글 삭제
+// 댓글 삭제
 export const deleteComment = async ({ commentId }: DeleteCommentParams) => {
   const response = await axiosInstance.delete(
     apiRoutes.CommentsDetail(commentId)
