+++ src/api/users.ts
@@ -1,5 +1,6 @@
 import axiosInstance from "./axiosInstance";
 import { apiRoutes } from "./apiRoutes";
+import { UpdateUser, UserMeImage } from "@/types/users";
 
 export interface UserResponse {
   id: number;
@@ -12,17 +13,13 @@ export interface UserResponse {
   userId?: number;
 }
 
-export const getUserInfo = async () => {
-  const response = await axiosInstance.get<UserResponse>(apiRoutes.usersMe());
-  return response.data;
-};
-
 interface SignUpRequest {
   email: string;
   nickname: string;
   password: string;
 }
 
+// 회원가입 (POST)
 export const signUp = async ({
   payload,
 }: {
@@ -33,5 +30,33 @@ export const signUp = async ({
     apiRoutes.users(),
     payload
   );
+  return response.data;
+};
+
+// 내 정보 조회 (GET)
+export const getUserInfo = async () => {
+  const response = await axiosInstance.get<UserResponse>(apiRoutes.usersMe());
+  return response.data;
+};
+
+// 내 정보 수정 (PUT)
+export const updateProfile = async (data: UpdateUser) => {
+  const res = await axiosInstance.put(apiRoutes.usersMe(), data, {
+    headers: {
+      "Content-Type": "application/json",
+    },
+  });
+  return res.data;
+};
+
+// 프로필 이미지 업로드 (POST)
+export const uploadProfileImage = async (
+  formData: FormData
+): Promise<UserMeImage> => {
+  const response = await axiosInstance.post(apiRoutes.userMeImage(), formData, {
+    headers: {
+      "Content-Type": "multipart/form-data",
+    },
+  });
   return response.data;
 };
